Foliage Analysis and comparative results between LiDAR and DTRF
This submission includes the following files:
CSV files containing the DTRF and LiDAR presence of the foliage/vegetation among the three regions.
notebook performing results analysis and comparisons among the DTRF and LiDAR data among the three regions and drawing Venn diagrams showing the overlap of the foliage between LiDAR and DTRF data.